requirements/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cryptography==36.0.0
defusedxml==0.7.1
dj-rest-auth==2.1.11
Django==3.2.10
django-allauth==0.47.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
idna==3.3
mysqlclient==2.1.0
oauthlib==3.1.1
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
sqlparse==0.4.2
typing_extensions==4.0.1
urllib3==1.26.7
gunicorn==20.1.0
Pillow==8.4.0

